Fixing broken build
IMPORTANT: Because of the way the tests are discovered on Linux, when
adding new empty tests to be filled out later, comment out the bodies
and not the tests themselves. If you comment the tests, they will still
be discovered and added to the list of tests to be run.