TosLang: Adding activation record abstraction
An activation record contains information about the total memory size of
a function's parameters and about the function's local variable. This
information will be useful later on during code generation when we'll do
the memory layout. The thing to understand about activation records is
that they're basically the compile time equivalent of the run time
structure called the call stack frame.